Extract assistant advisory draft attachment
