+++ control-plane/aegisops_control_plane/assistant_context.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from dataclasses import replace
 import re
 from typing import Any, Callable, Mapping
 
@@ -223,6 +224,31 @@ def _advisory_text_claims_authority_or_scope_expansion(text: object) -> tuple[st
         flags.append("scope_expansion_attempt")
 
     return _dedupe_strings(tuple(flags))
+
+
+def _assistant_advisory_draft_without_revision_history(
+    draft: Mapping[str, object],
+) -> dict[str, object]:
+    return {
+        str(key): value
+        for key, value in draft.items()
+        if str(key) != "revision_history"
+    }
+
+
+def _assistant_advisory_draft_revision_history(
+    draft: Mapping[str, object],
+) -> tuple[dict[str, object], ...]:
+    raw_history = draft.get("revision_history", ())
+    if not isinstance(raw_history, (list, tuple)):
+        return ()
+    revision_history: list[dict[str, object]] = []
+    for entry in raw_history:
+        if isinstance(entry, Mapping):
+            revision_history.append(
+                _assistant_advisory_draft_without_revision_history(entry)
+            )
+    return tuple(revision_history)
 
 
 def _build_assistant_advisory_output(
@@ -905,3 +931,61 @@ class AssistantContextAssembler:
         context_snapshot = self.inspect_assistant_context(record_family, record_id)
         self._service._require_reviewed_case_scoped_advisory_read(context_snapshot)
         return self._recommendation_draft_snapshot_from_context(context_snapshot)
+
+    def attach_assistant_advisory_draft(
+        self,
+        record_family: str,
+        record_id: str,
+    ) -> RecommendationRecord | AITraceRecord:
+        record_family = self._service._require_non_empty_string(
+            record_family,
+            "record_family",
+        )
+        record_id = self._service._require_non_empty_string(record_id, "record_id")
+        if record_family not in {"recommendation", "ai_trace"}:
+            raise ValueError(
+                "assistant advisory drafts may only be attached to "
+                "'recommendation' or 'ai_trace' records"
+            )
+
+        record_type = self._record_types_by_family[record_family]
+        record = self._service._store.get(record_type, record_id)
+        if record is None:
+            raise LookupError(
+                f"Missing {record_family} record {record_id!r} for advisory draft attachment"
+            )
+        if not isinstance(record, (RecommendationRecord, AITraceRecord)):
+            raise TypeError(
+                "assistant advisory drafts may only be attached to recommendation "
+                "or ai_trace records"
+            )
+
+        draft_snapshot = self.render_recommendation_draft(record_family, record_id)
+        attached_draft = {
+            "draft_id": f"assistant-advisory-draft:{record_family}:{record_id}",
+            "source_record_family": record_family,
+            "source_record_id": record_id,
+            "review_lifecycle_state": record.lifecycle_state,
+            **draft_snapshot.recommendation_draft,
+            "linked_alert_ids": draft_snapshot.linked_alert_ids,
+            "linked_case_ids": draft_snapshot.linked_case_ids,
+            "linked_evidence_ids": draft_snapshot.linked_evidence_ids,
+            "linked_recommendation_ids": draft_snapshot.linked_recommendation_ids,
+            "linked_reconciliation_ids": draft_snapshot.linked_reconciliation_ids,
+        }
+        current_attached_draft = _assistant_advisory_draft_without_revision_history(
+            record.assistant_advisory_draft
+        )
+        if current_attached_draft == attached_draft:
+            return record
+        revision_history = _assistant_advisory_draft_revision_history(
+            record.assistant_advisory_draft
+        )
+        if current_attached_draft:
+            attached_draft["revision_history"] = (
+                *revision_history,
+                current_attached_draft,
+            )
+        return self._service.persist_record(
+            replace(record, assistant_advisory_draft=attached_draft)
+        )
